gen/nl/lxtreme/libtdl/grammar/adv/AdvTdlParser.java
// Generated from AdvTdl.g4 by ANTLR 4.2.2

/*
 * LibTDL - Library for parsing/handling the "Trigger Definition Language".
 *
 * (C) Copyright 2012-2013 - J.W. Janssen, <j.w.janssen@lxtreme.nl>
 *
 * Licensed under Apache Software License version 2.0, see <http://www.apache.org/licenses/LICENSE-2.0.html>.
 */
package nl.lxtreme.libtdl.grammar.adv;


import org.antlr.v4.runtime.atn.*;
import org.antlr.v4.runtime.dfa.DFA;
import org.antlr.v4.runtime.*;
import org.antlr.v4.runtime.misc.*;
import org.antlr.v4.runtime.tree.*;
import java.util.List;
import java.util.Iterator;
import java.util.ArrayList;

@SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast"})
public class AdvTdlParser extends Parser {
	protected static final DFA[] _decisionToDFA;
	protected static final PredictionContextCache _sharedContextCache =
		new PredictionContextCache();
	public static final int
		T__0=1, COMMENT=2, NL=3, WS=4, ASSIGN=5, EQUALS_TO=6, MASK=7, VALUE=8, 
		RISING=9, FALLING=10, BOTH=11, NEITHER=12, STAGE=13, CAPTURE=14, NOP=15, 
		ANY=16, WHEN=17, OCCURS=18, START=19, STOP=20, CLEAR=21, GOTO=22, NEXT=23, 
		ELSE=24, ON=25, COMMA=26, COLON=27, NOT=28, XOR=29, AND=30, OR=31, LPAREN=32, 
		RPAREN=33, BIN_LITERAL=34, HEX_LITERAL=35, OCT_LITERAL=36, DEC_LITERAL=37, 
		TIME_UNIT=38, TERM_NAME=39, TIMER_NAME=40, RANGE_NAME=41, EDGE_NAME=42;
	public static final String[] tokenNames = {
		"<INVALID>", "'..'", "COMMENT", "NL", "WS", "ASSIGN", "'='", "'mask'", 
		"'value'", "'rising'", "'falling'", "'both'", "'neither'", "'stage'", 
		"'capture'", "'nop'", "'any'", "'when'", "'occurs'", "'start'", "'stop'", 
		"'clear'", "'goto'", "'next'", "'else'", "'on'", "','", "':'", "'~'", 
		"'^'", "'&'", "'|'", "'('", "')'", "BIN_LITERAL", "HEX_LITERAL", "OCT_LITERAL", 
		"DEC_LITERAL", "TIME_UNIT", "TERM_NAME", "TIMER_NAME", "RANGE_NAME", "EDGE_NAME"
	};
	public static final int
		RULE_prog = 0, RULE_decl = 1, RULE_termDecl = 2, RULE_timerDecl = 3, RULE_rangeDecl = 4, 
		RULE_edgeTermDecl = 5, RULE_edgeDecl = 6, RULE_stageDef = 7, RULE_whenAction = 8, 
		RULE_elseAction = 9, RULE_termExpr = 10, RULE_expr = 11, RULE_number = 12, 
		RULE_decNumber = 13;
	public static final String[] ruleNames = {
		"prog", "decl", "termDecl", "timerDecl", "rangeDecl", "edgeTermDecl", 
		"edgeDecl", "stageDef", "whenAction", "elseAction", "termExpr", "expr", 
		"number", "decNumber"
	};

	@Override
	public String getGrammarFileName() { return "AdvTdl.g4"; }

	@Override
	public String[] getTokenNames() { return tokenNames; }

	@Override
	public String[] getRuleNames() { return ruleNames; }

	@Override
	public String getSerializedATN() { return _serializedATN; }

	@Override
	public ATN getATN() { return _ATN; }

	public AdvTdlParser(TokenStream input) {
		super(input);
		_interp = new ParserATNSimulator(this,_ATN,_decisionToDFA,_sharedContextCache);
	}
	public static class ProgContext extends ParserRuleContext {
		public DeclContext decl(int i) {
			return getRuleContext(DeclContext.class,i);
		}
		public TerminalNode EOF() { return getToken(AdvTdlParser.EOF, 0); }
		public List<DeclContext> decl() {
			return getRuleContexts(DeclContext.class);
		}
		public List<StageDefContext> stageDef() {
			return getRuleContexts(StageDefContext.class);
		}
		public StageDefContext stageDef(int i) {
			return getRuleContext(StageDefContext.class,i);
		}
		public ProgContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_prog; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitProg(this);
			else return visitor.visitChildren(this);
		}
	}

	public final ProgContext prog() throws RecognitionException {
		ProgContext _localctx = new ProgContext(_ctx, getState());
		enterRule(_localctx, 0, RULE_prog);
		int _la;
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(32);
			_errHandler.sync(this);
			_la = _input.LA(1);
			while ((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << STAGE) | (1L << TERM_NAME) | (1L << TIMER_NAME) | (1L << RANGE_NAME) | (1L << EDGE_NAME))) != 0)) {
				{
				setState(30);
				switch (_input.LA(1)) {
				case TERM_NAME:
				case TIMER_NAME:
				case RANGE_NAME:
				case EDGE_NAME:
					{
					setState(28); decl();
					}
					break;
				case STAGE:
					{
					setState(29); stageDef();
					}
					break;
				default:
					throw new NoViableAltException(this);
				}
				}
				setState(34);
				_errHandler.sync(this);
				_la = _input.LA(1);
			}
			setState(35); match(EOF);
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class DeclContext extends ParserRuleContext {
		public TerminalNode WS(int i) {
			return getToken(AdvTdlParser.WS, i);
		}
		public TermDeclContext termDecl() {
			return getRuleContext(TermDeclContext.class,0);
		}
		public List<TerminalNode> WS() { return getTokens(AdvTdlParser.WS); }
		public EdgeDeclContext edgeDecl() {
			return getRuleContext(EdgeDeclContext.class,0);
		}
		public TimerDeclContext timerDecl() {
			return getRuleContext(TimerDeclContext.class,0);
		}
		public RangeDeclContext rangeDecl() {
			return getRuleContext(RangeDeclContext.class,0);
		}
		public DeclContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_decl; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitDecl(this);
			else return visitor.visitChildren(this);
		}
	}

	public final DeclContext decl() throws RecognitionException {
		DeclContext _localctx = new DeclContext(_ctx, getState());
		enterRule(_localctx, 2, RULE_decl);
		int _la;
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(41);
			switch (_input.LA(1)) {
			case TERM_NAME:
				{
				setState(37); termDecl();
				}
				break;
			case TIMER_NAME:
				{
				setState(38); timerDecl();
				}
				break;
			case RANGE_NAME:
				{
				setState(39); rangeDecl();
				}
				break;
			case EDGE_NAME:
				{
				setState(40); edgeDecl();
				}
				break;
			default:
				throw new NoViableAltException(this);
			}
			setState(46);
			_errHandler.sync(this);
			_la = _input.LA(1);
			while (_la==WS) {
				{
				{
				setState(43); match(WS);
				}
				}
				setState(48);
				_errHandler.sync(this);
				_la = _input.LA(1);
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class TermDeclContext extends ParserRuleContext {
		public Token name;
		public NumberContext mask;
		public NumberContext value;
		public TerminalNode ASSIGN() { return getToken(AdvTdlParser.ASSIGN, 0); }
		public List<NumberContext> number() {
			return getRuleContexts(NumberContext.class);
		}
		public List<TerminalNode> EQUALS_TO() { return getTokens(AdvTdlParser.EQUALS_TO); }
		public TerminalNode XOR() { return getToken(AdvTdlParser.XOR, 0); }
		public TerminalNode COMMA() { return getToken(AdvTdlParser.COMMA, 0); }
		public TerminalNode MASK() { return getToken(AdvTdlParser.MASK, 0); }
		public TerminalNode EQUALS_TO(int i) {
			return getToken(AdvTdlParser.EQUALS_TO, i);
		}
		public NumberContext number(int i) {
			return getRuleContext(NumberContext.class,i);
		}
		public TerminalNode VALUE() { return getToken(AdvTdlParser.VALUE, 0); }
		public TerminalNode TERM_NAME() { return getToken(AdvTdlParser.TERM_NAME, 0); }
		public TermDeclContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_termDecl; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitTermDecl(this);
			else return visitor.visitChildren(this);
		}
	}

	public final TermDeclContext termDecl() throws RecognitionException {
		TermDeclContext _localctx = new TermDeclContext(_ctx, getState());
		enterRule(_localctx, 4, RULE_termDecl);
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(49); ((TermDeclContext)_localctx).name = match(TERM_NAME);
			setState(50); match(ASSIGN);
			setState(74);
			switch ( getInterpreter().adaptivePredict(_input,4,_ctx) ) {
			case 1:
				{
				{
				setState(51); match(MASK);
				setState(52); match(EQUALS_TO);
				setState(53); ((TermDeclContext)_localctx).mask = number();
				}
				setState(55); match(COMMA);
				{
				setState(56); match(VALUE);
				setState(57); match(EQUALS_TO);
				setState(58); ((TermDeclContext)_localctx).value = number();
				}
				}
				break;

			case 2:
				{
				{
				setState(60); match(MASK);
				setState(61); match(EQUALS_TO);
				setState(62); number();
				}
				 notifyErrorListeners("missing term value"); 
				}
				break;

			case 3:
				{
				{
				setState(66); ((TermDeclContext)_localctx).mask = number();
				setState(67); match(XOR);
				setState(68); ((TermDeclContext)_localctx).value = number();
				}
				}
				break;

			case 4:
				{
				{
				setState(70); number();
				}
				 notifyErrorListeners("missing term value"); 
				}
				break;

			case 5:
				{
				 notifyErrorListeners("missing mask and value"); 
				}
				break;
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class TimerDeclContext extends ParserRuleContext {
		public Token name;
		public NumberContext value;
		public Token unit;
		public TerminalNode ASSIGN() { return getToken(AdvTdlParser.ASSIGN, 0); }
		public NumberContext number() {
			return getRuleContext(NumberContext.class,0);
		}
		public TerminalNode TIMER_NAME() { return getToken(AdvTdlParser.TIMER_NAME, 0); }
		public TerminalNode TIME_UNIT() { return getToken(AdvTdlParser.TIME_UNIT, 0); }
		public TimerDeclContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_timerDecl; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitTimerDecl(this);
			else return visitor.visitChildren(this);
		}
	}

	public final TimerDeclContext timerDecl() throws RecognitionException {
		TimerDeclContext _localctx = new TimerDeclContext(_ctx, getState());
		enterRule(_localctx, 6, RULE_timerDecl);
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(76); ((TimerDeclContext)_localctx).name = match(TIMER_NAME);
			setState(77); match(ASSIGN);
			setState(85);
			switch ( getInterpreter().adaptivePredict(_input,5,_ctx) ) {
			case 1:
				{
				setState(78); ((TimerDeclContext)_localctx).value = number();
				setState(79); ((TimerDeclContext)_localctx).unit = match(TIME_UNIT);
				}
				break;

			case 2:
				{
				setState(81); number();
				 notifyErrorListeners("missing time unit"); 
				}
				break;

			case 3:
				{
				 notifyErrorListeners("missing time value and unit"); 
				}
				break;
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class RangeDeclContext extends ParserRuleContext {
		public Token name;
		public NumberContext lowerBound;
		public NumberContext upperBound;
		public TerminalNode ASSIGN() { return getToken(AdvTdlParser.ASSIGN, 0); }
		public List<NumberContext> number() {
			return getRuleContexts(NumberContext.class);
		}
		public TerminalNode RANGE_NAME() { return getToken(AdvTdlParser.RANGE_NAME, 0); }
		public NumberContext number(int i) {
			return getRuleContext(NumberContext.class,i);
		}
		public RangeDeclContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_rangeDecl; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitRangeDecl(this);
			else return visitor.visitChildren(this);
		}
	}

	public final RangeDeclContext rangeDecl() throws RecognitionException {
		RangeDeclContext _localctx = new RangeDeclContext(_ctx, getState());
		enterRule(_localctx, 8, RULE_rangeDecl);
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(87); ((RangeDeclContext)_localctx).name = match(RANGE_NAME);
			setState(88); match(ASSIGN);
			setState(101);
			switch ( getInterpreter().adaptivePredict(_input,6,_ctx) ) {
			case 1:
				{
				setState(89); ((RangeDeclContext)_localctx).lowerBound = number();
				setState(90); match(1);
				setState(91); ((RangeDeclContext)_localctx).upperBound = number();
				}
				break;

			case 2:
				{
				setState(93); number();
				setState(94); match(1);
				 notifyErrorListeners("missing upper bound"); 
				}
				break;

			case 3:
				{
				setState(97); number();
				 notifyErrorListeners("missing upper bound"); 
				}
				break;

			case 4:
				{
				 notifyErrorListeners("invalid range definition, needs lower and upper bound"); 
				}
				break;
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class EdgeTermDeclContext extends ParserRuleContext {
		public Token name;
		public NumberContext mask;
		public NumberContext number() {
			return getRuleContext(NumberContext.class,0);
		}
		public TerminalNode EQUALS_TO() { return getToken(AdvTdlParser.EQUALS_TO, 0); }
		public TerminalNode BOTH() { return getToken(AdvTdlParser.BOTH, 0); }
		public TerminalNode NEITHER() { return getToken(AdvTdlParser.NEITHER, 0); }
		public TerminalNode RISING() { return getToken(AdvTdlParser.RISING, 0); }
		public TerminalNode FALLING() { return getToken(AdvTdlParser.FALLING, 0); }
		public EdgeTermDeclContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_edgeTermDecl; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitEdgeTermDecl(this);
			else return visitor.visitChildren(this);
		}
	}

	public final EdgeTermDeclContext edgeTermDecl() throws RecognitionException {
		EdgeTermDeclContext _localctx = new EdgeTermDeclContext(_ctx, getState());
		enterRule(_localctx, 10, RULE_edgeTermDecl);
		int _la;
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(103);
			((EdgeTermDeclContext)_localctx).name = _input.LT(1);
			_la = _input.LA(1);
			if ( !((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << RISING) | (1L << FALLING) | (1L << BOTH) | (1L << NEITHER))) != 0)) ) {
				((EdgeTermDeclContext)_localctx).name = (Token)_errHandler.recoverInline(this);
			}
			consume();
			setState(104); match(EQUALS_TO);
			setState(107);
			switch (_input.LA(1)) {
			case BIN_LITERAL:
			case HEX_LITERAL:
			case OCT_LITERAL:
			case DEC_LITERAL:
				{
				setState(105); ((EdgeTermDeclContext)_localctx).mask = number();
				}
				break;
			case EOF:
			case WS:
			case STAGE:
			case COMMA:
			case TERM_NAME:
			case TIMER_NAME:
			case RANGE_NAME:
			case EDGE_NAME:
				{
				 notifyErrorListeners("missing edge value"); 
				}
				break;
			default:
				throw new NoViableAltException(this);
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class EdgeDeclContext extends ParserRuleContext {
		public Token name;
		public EdgeTermDeclContext edgeTermDecl;
		public List<EdgeTermDeclContext> terms = new ArrayList<EdgeTermDeclContext>();
		public TerminalNode ASSIGN() { return getToken(AdvTdlParser.ASSIGN, 0); }
		public List<EdgeTermDeclContext> edgeTermDecl() {
			return getRuleContexts(EdgeTermDeclContext.class);
		}
		public TerminalNode EDGE_NAME() { return getToken(AdvTdlParser.EDGE_NAME, 0); }
		public List<TerminalNode> COMMA() { return getTokens(AdvTdlParser.COMMA); }
		public EdgeTermDeclContext edgeTermDecl(int i) {
			return getRuleContext(EdgeTermDeclContext.class,i);
		}
		public TerminalNode COMMA(int i) {
			return getToken(AdvTdlParser.COMMA, i);
		}
		public EdgeDeclContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_edgeDecl; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitEdgeDecl(this);
			else return visitor.visitChildren(this);
		}
	}

	public final EdgeDeclContext edgeDecl() throws RecognitionException {
		EdgeDeclContext _localctx = new EdgeDeclContext(_ctx, getState());
		enterRule(_localctx, 12, RULE_edgeDecl);
		int _la;
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(109); ((EdgeDeclContext)_localctx).name = match(EDGE_NAME);
			setState(110); match(ASSIGN);
			setState(120);
			switch (_input.LA(1)) {
			case RISING:
			case FALLING:
			case BOTH:
			case NEITHER:
				{
				{
				setState(111); ((EdgeDeclContext)_localctx).edgeTermDecl = edgeTermDecl();
				((EdgeDeclContext)_localctx).terms.add(((EdgeDeclContext)_localctx).edgeTermDecl);
				setState(116);
				_errHandler.sync(this);
				_la = _input.LA(1);
				while (_la==COMMA) {
					{
					{
					setState(112); match(COMMA);
					setState(113); ((EdgeDeclContext)_localctx).edgeTermDecl = edgeTermDecl();
					((EdgeDeclContext)_localctx).terms.add(((EdgeDeclContext)_localctx).edgeTermDecl);
					}
					}
					setState(118);
					_errHandler.sync(this);
					_la = _input.LA(1);
				}
				}
				}
				break;
			case EOF:
			case WS:
			case STAGE:
			case TERM_NAME:
			case TIMER_NAME:
			case RANGE_NAME:
			case EDGE_NAME:
				{
				 notifyErrorListeners("invalid edge definition, needs at least one edge term declaration"); 
				}
				break;
			default:
				throw new NoViableAltException(this);
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class StageDefContext extends ParserRuleContext {
		public DecNumberContext n;
		public TermExprContext captureExpr;
		public TermExprContext ifExpr;
		public DecNumberContext occurrence;
		public TermExprContext elseExpr;
		public TerminalNode CAPTURE() { return getToken(AdvTdlParser.CAPTURE, 0); }
		public List<TermExprContext> termExpr() {
			return getRuleContexts(TermExprContext.class);
		}
		public TerminalNode ELSE() { return getToken(AdvTdlParser.ELSE, 0); }
		public WhenActionContext whenAction() {
			return getRuleContext(WhenActionContext.class,0);
		}
		public TerminalNode ON() { return getToken(AdvTdlParser.ON, 0); }
		public TerminalNode COLON() { return getToken(AdvTdlParser.COLON, 0); }
		public TerminalNode OCCURS() { return getToken(AdvTdlParser.OCCURS, 0); }
		public List<DecNumberContext> decNumber() {
			return getRuleContexts(DecNumberContext.class);
		}
		public ElseActionContext elseAction() {
			return getRuleContext(ElseActionContext.class,0);
		}
		public TerminalNode STAGE() { return getToken(AdvTdlParser.STAGE, 0); }
		public DecNumberContext decNumber(int i) {
			return getRuleContext(DecNumberContext.class,i);
		}
		public TermExprContext termExpr(int i) {
			return getRuleContext(TermExprContext.class,i);
		}
		public TerminalNode WHEN() { return getToken(AdvTdlParser.WHEN, 0); }
		public StageDefContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_stageDef; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitStageDef(this);
			else return visitor.visitChildren(this);
		}
	}

	public final StageDefContext stageDef() throws RecognitionException {
		StageDefContext _localctx = new StageDefContext(_ctx, getState());
		enterRule(_localctx, 14, RULE_stageDef);
		int _la;
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(132);
			switch ( getInterpreter().adaptivePredict(_input,10,_ctx) ) {
			case 1:
				{
				setState(122); match(STAGE);
				setState(123); ((StageDefContext)_localctx).n = decNumber();
				setState(124); match(COLON);
				}
				break;

			case 2:
				{
				setState(126); match(STAGE);
				setState(127); decNumber();
				 notifyErrorListeners("missing colon"); 
				}
				break;

			case 3:
				{
				setState(130); match(STAGE);
				 notifyErrorListeners("missing stage ID"); 
				}
				break;
			}
			setState(139);
			switch ( getInterpreter().adaptivePredict(_input,11,_ctx) ) {
			case 1:
				{
				setState(134); match(CAPTURE);
				setState(135); ((StageDefContext)_localctx).captureExpr = termExpr();
				}
				break;

			case 2:
				{
				setState(136); match(CAPTURE);
				 notifyErrorListeners("missing capture expression"); 
				}
				break;

			case 3:
				{
				 notifyErrorListeners("missing capture clause"); 
				}
				break;
			}
			setState(164);
			switch ( getInterpreter().adaptivePredict(_input,15,_ctx) ) {
			case 1:
				{
				setState(141); match(WHEN);
				setState(142); ((StageDefContext)_localctx).ifExpr = termExpr();
				setState(145);
				_la = _input.LA(1);
				if (_la==OCCURS) {
					{
					setState(143); match(OCCURS);
					setState(144); ((StageDefContext)_localctx).occurrence = decNumber();
					}
				}

				setState(147); whenAction();
				}
				break;

			case 2:
				{
				setState(149); match(WHEN);
				setState(150); termExpr();
				setState(153);
				_la = _input.LA(1);
				if (_la==OCCURS) {
					{
					setState(151); match(OCCURS);
					setState(152); decNumber();
					}
				}

				 notifyErrorListeners("missing when action"); 
				}
				break;

			case 3:
				{
				setState(157); match(WHEN);
				setState(160);
				_la = _input.LA(1);
				if (_la==OCCURS) {
					{
					setState(158); match(OCCURS);
					setState(159); decNumber();
					}
				}

				 notifyErrorListeners("missing when expression"); 
				}
				break;

			case 4:
				{
				 notifyErrorListeners("missing when clause"); 
				}
				break;
			}
			setState(182);
			switch ( getInterpreter().adaptivePredict(_input,16,_ctx) ) {
			case 1:
				{
				setState(166); match(ELSE);
				setState(167); match(ON);
				setState(168); ((StageDefContext)_localctx).elseExpr = termExpr();
				setState(169); elseAction();
				}
				break;

			case 2:
				{
				setState(171); match(ELSE);
				setState(172); match(ON);
				setState(173); termExpr();
				 notifyErrorListeners("missing else action"); 
				}
				break;

			case 3:
				{
				setState(176); match(ELSE);
				setState(177); match(ON);
				 notifyErrorListeners("missing else expression"); 
				}
				break;

			case 4:
				{
				setState(179); match(ELSE);
				 notifyErrorListeners("missing on"); 
				}
				break;

			case 5:
				{
				 notifyErrorListeners("missing else clause"); 
				}
				break;
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class WhenActionContext extends ParserRuleContext {
		public Token action;
		public Token name;
		public TerminalNode CAPTURE() { return getToken(AdvTdlParser.CAPTURE, 0); }
		public TerminalNode NEXT() { return getToken(AdvTdlParser.NEXT, 0); }
		public TerminalNode TIMER_NAME() { return getToken(AdvTdlParser.TIMER_NAME, 0); }
		public TerminalNode GOTO() { return getToken(AdvTdlParser.GOTO, 0); }
		public TerminalNode STOP() { return getToken(AdvTdlParser.STOP, 0); }
		public TerminalNode START() { return getToken(AdvTdlParser.START, 0); }
		public TerminalNode CLEAR() { return getToken(AdvTdlParser.CLEAR, 0); }
		public WhenActionContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_whenAction; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitWhenAction(this);
			else return visitor.visitChildren(this);
		}
	}

	public final WhenActionContext whenAction() throws RecognitionException {
		WhenActionContext _localctx = new WhenActionContext(_ctx, getState());
		enterRule(_localctx, 16, RULE_whenAction);
		int _la;
		try {
			setState(194);
			switch ( getInterpreter().adaptivePredict(_input,18,_ctx) ) {
			case 1:
				enterOuterAlt(_localctx, 1);
				{
				setState(184);
				((WhenActionContext)_localctx).action = _input.LT(1);
				_la = _input.LA(1);
				if ( !((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << START) | (1L << STOP) | (1L << CLEAR))) != 0)) ) {
					((WhenActionContext)_localctx).action = (Token)_errHandler.recoverInline(this);
				}
				consume();
				setState(185); ((WhenActionContext)_localctx).name = match(TIMER_NAME);
				}
				break;

			case 2:
				enterOuterAlt(_localctx, 2);
				{
				setState(186);
				((WhenActionContext)_localctx).action = _input.LT(1);
				_la = _input.LA(1);
				if ( !(_la==START || _la==STOP) ) {
					((WhenActionContext)_localctx).action = (Token)_errHandler.recoverInline(this);
				}
				consume();
				setState(187); ((WhenActionContext)_localctx).name = match(CAPTURE);
				}
				break;

			case 3:
				enterOuterAlt(_localctx, 3);
				{
				setState(192);
				switch ( getInterpreter().adaptivePredict(_input,17,_ctx) ) {
				case 1:
					{
					setState(188); match(GOTO);
					setState(189); match(NEXT);
					}
					break;

				case 2:
					{
					setState(190); match(GOTO);
					 notifyErrorListeners("missing next"); 
					}
					break;
				}
				}
				break;
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class ElseActionContext extends ParserRuleContext {
		public DecNumberContext n;
		public TerminalNode GOTO() { return getToken(AdvTdlParser.GOTO, 0); }
		public DecNumberContext decNumber() {
			return getRuleContext(DecNumberContext.class,0);
		}
		public ElseActionContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_elseAction; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitElseAction(this);
			else return visitor.visitChildren(this);
		}
	}

	public final ElseActionContext elseAction() throws RecognitionException {
		ElseActionContext _localctx = new ElseActionContext(_ctx, getState());
		enterRule(_localctx, 18, RULE_elseAction);
		try {
			setState(200);
			switch ( getInterpreter().adaptivePredict(_input,19,_ctx) ) {
			case 1:
				enterOuterAlt(_localctx, 1);
				{
				setState(196); match(GOTO);
				setState(197); ((ElseActionContext)_localctx).n = decNumber();
				}
				break;

			case 2:
				enterOuterAlt(_localctx, 2);
				{
				setState(198); match(GOTO);
				 notifyErrorListeners("missing level ID"); 
				}
				break;
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class TermExprContext extends ParserRuleContext {
		public Token e;
		public ExprContext expr() {
			return getRuleContext(ExprContext.class,0);
		}
		public TerminalNode NOP() { return getToken(AdvTdlParser.NOP, 0); }
		public TerminalNode ANY() { return getToken(AdvTdlParser.ANY, 0); }
		public TermExprContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_termExpr; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitTermExpr(this);
			else return visitor.visitChildren(this);
		}
	}

	public final TermExprContext termExpr() throws RecognitionException {
		TermExprContext _localctx = new TermExprContext(_ctx, getState());
		enterRule(_localctx, 20, RULE_termExpr);
		try {
			setState(205);
			switch (_input.LA(1)) {
			case NOP:
				enterOuterAlt(_localctx, 1);
				{
				setState(202); ((TermExprContext)_localctx).e = match(NOP);
				}
				break;
			case ANY:
				enterOuterAlt(_localctx, 2);
				{
				setState(203); ((TermExprContext)_localctx).e = match(ANY);
				}
				break;
			case NOT:
			case LPAREN:
			case TERM_NAME:
			case TIMER_NAME:
			case RANGE_NAME:
			case EDGE_NAME:
				enterOuterAlt(_localctx, 3);
				{
				setState(204); expr(0);
				}
				break;
			default:
				throw new NoViableAltException(this);
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class ExprContext extends ParserRuleContext {
		public ExprContext lhs;
		public Token op;
		public ExprContext rhs;
		public Token term;
		public TerminalNode NOT() { return getToken(AdvTdlParser.NOT, 0); }
		public List<ExprContext> expr() {
			return getRuleContexts(ExprContext.class);
		}
		public TerminalNode RANGE_NAME() { return getToken(AdvTdlParser.RANGE_NAME, 0); }
		public TerminalNode XOR() { return getToken(AdvTdlParser.XOR, 0); }
		public ExprContext expr(int i) {
			return getRuleContext(ExprContext.class,i);
		}
		public TerminalNode LPAREN() { return getToken(AdvTdlParser.LPAREN, 0); }
		public TerminalNode EDGE_NAME() { return getToken(AdvTdlParser.EDGE_NAME, 0); }
		public TerminalNode TIMER_NAME() { return getToken(AdvTdlParser.TIMER_NAME, 0); }
		public TerminalNode AND() { return getToken(AdvTdlParser.AND, 0); }
		public TerminalNode RPAREN() { return getToken(AdvTdlParser.RPAREN, 0); }
		public TerminalNode OR() { return getToken(AdvTdlParser.OR, 0); }
		public TerminalNode TERM_NAME() { return getToken(AdvTdlParser.TERM_NAME, 0); }
		public ExprContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_expr; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitExpr(this);
			else return visitor.visitChildren(this);
		}
	}

	public final ExprContext expr() throws RecognitionException {
		return expr(0);
	}

	private ExprContext expr(int _p) throws RecognitionException {
		ParserRuleContext _parentctx = _ctx;
		int _parentState = getState();
		ExprContext _localctx = new ExprContext(_ctx, _parentState);
		ExprContext _prevctx = _localctx;
		int _startState = 22;
		enterRecursionRule(_localctx, 22, RULE_expr, _p);
		int _la;
		try {
			int _alt;
			enterOuterAlt(_localctx, 1);
			{
			setState(215);
			switch (_input.LA(1)) {
			case NOT:
				{
				setState(208); ((ExprContext)_localctx).op = match(NOT);
				setState(209); ((ExprContext)_localctx).rhs = expr(5);
				}
				break;
			case LPAREN:
				{
				setState(210); ((ExprContext)_localctx).op = match(LPAREN);
				setState(211); ((ExprContext)_localctx).lhs = expr(0);
				setState(212); match(RPAREN);
				}
				break;
			case TERM_NAME:
			case TIMER_NAME:
			case RANGE_NAME:
			case EDGE_NAME:
				{
				setState(214);
				((ExprContext)_localctx).term = _input.LT(1);
				_la = _input.LA(1);
				if ( !((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << TERM_NAME) | (1L << TIMER_NAME) | (1L << RANGE_NAME) | (1L << EDGE_NAME))) != 0)) ) {
					((ExprContext)_localctx).term = (Token)_errHandler.recoverInline(this);
				}
				consume();
				}
				break;
			default:
				throw new NoViableAltException(this);
			}
			_ctx.stop = _input.LT(-1);
			setState(228);
			_errHandler.sync(this);
			_alt = getInterpreter().adaptivePredict(_input,23,_ctx);
			while ( _alt!=2 && _alt!=ATN.INVALID_ALT_NUMBER ) {
				if ( _alt==1 ) {
					if ( _parseListeners!=null ) triggerExitRuleEvent();
					_prevctx = _localctx;
					{
					setState(226);
					switch ( getInterpreter().adaptivePredict(_input,22,_ctx) ) {
					case 1:
						{
						_localctx = new ExprContext(_parentctx, _parentState);
						_localctx.lhs = _prevctx;
						pushNewRecursionContext(_localctx, _startState, RULE_expr);
						setState(217);
						if (!(precpred(_ctx, 4))) throw new FailedPredicateException(this, "precpred(_ctx, 4)");
						setState(218); ((ExprContext)_localctx).op = match(AND);
						setState(219); ((ExprContext)_localctx).rhs = expr(5);
						}
						break;

					case 2:
						{
						_localctx = new ExprContext(_parentctx, _parentState);
						_localctx.lhs = _prevctx;
						pushNewRecursionContext(_localctx, _startState, RULE_expr);
						setState(220);
						if (!(precpred(_ctx, 3))) throw new FailedPredicateException(this, "precpred(_ctx, 3)");
						setState(221); ((ExprContext)_localctx).op = match(XOR);
						setState(222); ((ExprContext)_localctx).rhs = expr(4);
						}
						break;

					case 3:
						{
						_localctx = new ExprContext(_parentctx, _parentState);
						_localctx.lhs = _prevctx;
						pushNewRecursionContext(_localctx, _startState, RULE_expr);
						setState(223);
						if (!(precpred(_ctx, 2))) throw new FailedPredicateException(this, "precpred(_ctx, 2)");
						setState(224); ((ExprContext)_localctx).op = match(OR);
						setState(225); ((ExprContext)_localctx).rhs = expr(3);
						}
						break;
					}
					} 
				}
				setState(230);
				_errHandler.sync(this);
				_alt = getInterpreter().adaptivePredict(_input,23,_ctx);
			}
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			unrollRecursionContexts(_parentctx);
		}
		return _localctx;
	}

	public static class NumberContext extends ParserRuleContext {
		public TerminalNode HEX_LITERAL() { return getToken(AdvTdlParser.HEX_LITERAL, 0); }
		public TerminalNode BIN_LITERAL() { return getToken(AdvTdlParser.BIN_LITERAL, 0); }
		public TerminalNode DEC_LITERAL() { return getToken(AdvTdlParser.DEC_LITERAL, 0); }
		public TerminalNode OCT_LITERAL() { return getToken(AdvTdlParser.OCT_LITERAL, 0); }
		public NumberContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_number; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitNumber(this);
			else return visitor.visitChildren(this);
		}
	}

	public final NumberContext number() throws RecognitionException {
		NumberContext _localctx = new NumberContext(_ctx, getState());
		enterRule(_localctx, 24, RULE_number);
		int _la;
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(231);
			_la = _input.LA(1);
			if ( !((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << BIN_LITERAL) | (1L << HEX_LITERAL) | (1L << OCT_LITERAL) | (1L << DEC_LITERAL))) != 0)) ) {
			_errHandler.recoverInline(this);
			}
			consume();
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public static class DecNumberContext extends ParserRuleContext {
		public TerminalNode DEC_LITERAL() { return getToken(AdvTdlParser.DEC_LITERAL, 0); }
		public DecNumberContext(ParserRuleContext parent, int invokingState) {
			super(parent, invokingState);
		}
		@Override public int getRuleIndex() { return RULE_decNumber; }
		@Override
		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
			if ( visitor instanceof AdvTdlVisitor ) return ((AdvTdlVisitor<? extends T>)visitor).visitDecNumber(this);
			else return visitor.visitChildren(this);
		}
	}

	public final DecNumberContext decNumber() throws RecognitionException {
		DecNumberContext _localctx = new DecNumberContext(_ctx, getState());
		enterRule(_localctx, 26, RULE_decNumber);
		try {
			enterOuterAlt(_localctx, 1);
			{
			setState(233); match(DEC_LITERAL);
			}
		}
		catch (RecognitionException re) {
			_localctx.exception = re;
			_errHandler.reportError(this, re);
			_errHandler.recover(this, re);
		}
		finally {
			exitRule();
		}
		return _localctx;
	}

	public boolean sempred(RuleContext _localctx, int ruleIndex, int predIndex) {
		switch (ruleIndex) {
		case 11: return expr_sempred((ExprContext)_localctx, predIndex);
		}
		return true;
	}
	private boolean expr_sempred(ExprContext _localctx, int predIndex) {
		switch (predIndex) {
		case 0: return precpred(_ctx, 4);

		case 1: return precpred(_ctx, 3);

		case 2: return precpred(_ctx, 2);
		}
		return true;
	}

	public static final String _serializedATN =
		"\3\u0430\ud6d1\u8206\uad2d\u4417\uaef1\u8d80\uaadd\3,\u00ee\4\2\t\2\4"+
		"\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t"+
		"\13\4\f\t\f\4\r\t\r\4\16\t\16\4\17\t\17\3\2\3\2\7\2!\n\2\f\2\16\2$\13"+
		"\2\3\2\3\2\3\3\3\3\3\3\3\3\5\3,\n\3\3\3\7\3/\n\3\f\3\16\3\62\13\3\3\4"+
		"\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3\4\3"+
		"\4\3\4\3\4\3\4\3\4\3\4\3\4\5\4M\n\4\3\5\3\5\3\5\3\5\3\5\3\5\3\5\3\5\3"+
		"\5\5\5X\n\5\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\5"+
		"\6h\n\6\3\7\3\7\3\7\3\7\5\7n\n\7\3\b\3\b\3\b\3\b\3\b\7\bu\n\b\f\b\16\b"+
		"x\13\b\3\b\5\b{\n\b\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\5\t\u0087"+
		"\n\t\3\t\3\t\3\t\3\t\3\t\5\t\u008e\n\t\3\t\3\t\3\t\3\t\5\t\u0094\n\t\3"+
		"\t\3\t\3\t\3\t\3\t\3\t\5\t\u009c\n\t\3\t\3\t\3\t\3\t\3\t\5\t\u00a3\n\t"+
		"\3\t\3\t\5\t\u00a7\n\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t"+
		"\3\t\3\t\3\t\3\t\5\t\u00b9\n\t\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\5\n\u00c3"+
		"\n\n\5\n\u00c5\n\n\3\13\3\13\3\13\3\13\5\13\u00cb\n\13\3\f\3\f\3\f\5\f"+
		"\u00d0\n\f\3\r\3\r\3\r\3\r\3\r\3\r\3\r\3\r\5\r\u00da\n\r\3\r\3\r\3\r\3"+
		"\r\3\r\3\r\3\r\3\r\3\r\7\r\u00e5\n\r\f\r\16\r\u00e8\13\r\3\16\3\16\3\17"+
		"\3\17\3\17\2\3\30\20\2\4\6\b\n\f\16\20\22\24\26\30\32\34\2\7\3\2\13\16"+
		"\3\2\25\27\3\2\25\26\3\2),\3\2$\'\u010a\2\"\3\2\2\2\4+\3\2\2\2\6\63\3"+
		"\2\2\2\bN\3\2\2\2\nY\3\2\2\2\fi\3\2\2\2\16o\3\2\2\2\20\u0086\3\2\2\2\22"+
		"\u00c4\3\2\2\2\24\u00ca\3\2\2\2\26\u00cf\3\2\2\2\30\u00d9\3\2\2\2\32\u00e9"+
		"\3\2\2\2\34\u00eb\3\2\2\2\36!\5\4\3\2\37!\5\20\t\2 \36\3\2\2\2 \37\3\2"+
		"\2\2!$\3\2\2\2\" \3\2\2\2\"#\3\2\2\2#%\3\2\2\2$\"\3\2\2\2%&\7\2\2\3&\3"+
		"\3\2\2\2\',\5\6\4\2(,\5\b\5\2),\5\n\6\2*,\5\16\b\2+\'\3\2\2\2+(\3\2\2"+
		"\2+)\3\2\2\2+*\3\2\2\2,\60\3\2\2\2-/\7\6\2\2.-\3\2\2\2/\62\3\2\2\2\60"+
		".\3\2\2\2\60\61\3\2\2\2\61\5\3\2\2\2\62\60\3\2\2\2\63\64\7)\2\2\64L\7"+
		"\7\2\2\65\66\7\t\2\2\66\67\7\b\2\2\678\5\32\16\289\3\2\2\29:\7\34\2\2"+
		":;\7\n\2\2;<\7\b\2\2<=\5\32\16\2=M\3\2\2\2>?\7\t\2\2?@\7\b\2\2@A\5\32"+
		"\16\2AB\3\2\2\2BC\b\4\1\2CM\3\2\2\2DE\5\32\16\2EF\7\37\2\2FG\5\32\16\2"+
		"GM\3\2\2\2HI\5\32\16\2IJ\b\4\1\2JM\3\2\2\2KM\b\4\1\2L\65\3\2\2\2L>\3\2"+
		"\2\2LD\3\2\2\2LH\3\2\2\2LK\3\2\2\2M\7\3\2\2\2NO\7*\2\2OW\7\7\2\2PQ\5\32"+
		"\16\2QR\7(\2\2RX\3\2\2\2ST\5\32\16\2TU\b\5\1\2UX\3\2\2\2VX\b\5\1\2WP\3"+
		"\2\2\2WS\3\2\2\2WV\3\2\2\2X\t\3\2\2\2YZ\7+\2\2Zg\7\7\2\2[\\\5\32\16\2"+
		"\\]\7\3\2\2]^\5\32\16\2^h\3\2\2\2_`\5\32\16\2`a\7\3\2\2ab\b\6\1\2bh\3"+
		"\2\2\2cd\5\32\16\2de\b\6\1\2eh\3\2\2\2fh\b\6\1\2g[\3\2\2\2g_\3\2\2\2g"+
		"c\3\2\2\2gf\3\2\2\2h\13\3\2\2\2ij\t\2\2\2jm\7\b\2\2kn\5\32\16\2ln\b\7"+
		"\1\2mk\3\2\2\2ml\3\2\2\2n\r\3\2\2\2op\7,\2\2pz\7\7\2\2qv\5\f\7\2rs\7\34"+
		"\2\2su\5\f\7\2tr\3\2\2\2ux\3\2\2\2vt\3\2\2\2vw\3\2\2\2w{\3\2\2\2xv\3\2"+
		"\2\2y{\b\b\1\2zq\3\2\2\2zy\3\2\2\2{\17\3\2\2\2|}\7\17\2\2}~\5\34\17\2"+
		"~\177\7\35\2\2\177\u0087\3\2\2\2\u0080\u0081\7\17\2\2\u0081\u0082\5\34"+
		"\17\2\u0082\u0083\b\t\1\2\u0083\u0087\3\2\2\2\u0084\u0085\7\17\2\2\u0085"+
		"\u0087\b\t\1\2\u0086|\3\2\2\2\u0086\u0080\3\2\2\2\u0086\u0084\3\2\2\2"+
		"\u0087\u008d\3\2\2\2\u0088\u0089\7\20\2\2\u0089\u008e\5\26\f\2\u008a\u008b"+
		"\7\20\2\2\u008b\u008e\b\t\1\2\u008c\u008e\b\t\1\2\u008d\u0088\3\2\2\2"+
		"\u008d\u008a\3\2\2\2\u008d\u008c\3\2\2\2\u008e\u00a6\3\2\2\2\u008f\u0090"+
		"\7\23\2\2\u0090\u0093\5\26\f\2\u0091\u0092\7\24\2\2\u0092\u0094\5\34\17"+
		"\2\u0093\u0091\3\2\2\2\u0093\u0094\3\2\2\2\u0094\u0095\3\2\2\2\u0095\u0096"+
		"\5\22\n\2\u0096\u00a7\3\2\2\2\u0097\u0098\7\23\2\2\u0098\u009b\5\26\f"+
		"\2\u0099\u009a\7\24\2\2\u009a\u009c\5\34\17\2\u009b\u0099\3\2\2\2\u009b"+
		"\u009c\3\2\2\2\u009c\u009d\3\2\2\2\u009d\u009e\b\t\1\2\u009e\u00a7\3\2"+
		"\2\2\u009f\u00a2\7\23\2\2\u00a0\u00a1\7\24\2\2\u00a1\u00a3\5\34\17\2\u00a2"+
		"\u00a0\3\2\2\2\u00a2\u00a3\3\2\2\2\u00a3\u00a4\3\2\2\2\u00a4\u00a7\b\t"+
		"\1\2\u00a5\u00a7\b\t\1\2\u00a6\u008f\3\2\2\2\u00a6\u0097\3\2\2\2\u00a6"+
		"\u009f\3\2\2\2\u00a6\u00a5\3\2\2\2\u00a7\u00b8\3\2\2\2\u00a8\u00a9\7\32"+
		"\2\2\u00a9\u00aa\7\33\2\2\u00aa\u00ab\5\26\f\2\u00ab\u00ac\5\24\13\2\u00ac"+
		"\u00b9\3\2\2\2\u00ad\u00ae\7\32\2\2\u00ae\u00af\7\33\2\2\u00af\u00b0\5"+
		"\26\f\2\u00b0\u00b1\b\t\1\2\u00b1\u00b9\3\2\2\2\u00b2\u00b3\7\32\2\2\u00b3"+
		"\u00b4\7\33\2\2\u00b4\u00b9\b\t\1\2\u00b5\u00b6\7\32\2\2\u00b6\u00b9\b"+
		"\t\1\2\u00b7\u00b9\b\t\1\2\u00b8\u00a8\3\2\2\2\u00b8\u00ad\3\2\2\2\u00b8"+
		"\u00b2\3\2\2\2\u00b8\u00b5\3\2\2\2\u00b8\u00b7\3\2\2\2\u00b9\21\3\2\2"+
		"\2\u00ba\u00bb\t\3\2\2\u00bb\u00c5\7*\2\2\u00bc\u00bd\t\4\2\2\u00bd\u00c5"+
		"\7\20\2\2\u00be\u00bf\7\30\2\2\u00bf\u00c3\7\31\2\2\u00c0\u00c1\7\30\2"+
		"\2\u00c1\u00c3\b\n\1\2\u00c2\u00be\3\2\2\2\u00c2\u00c0\3\2\2\2\u00c3\u00c5"+
		"\3\2\2\2\u00c4\u00ba\3\2\2\2\u00c4\u00bc\3\2\2\2\u00c4\u00c2\3\2\2\2\u00c5"+
		"\23\3\2\2\2\u00c6\u00c7\7\30\2\2\u00c7\u00cb\5\34\17\2\u00c8\u00c9\7\30"+
		"\2\2\u00c9\u00cb\b\13\1\2\u00ca\u00c6\3\2\2\2\u00ca\u00c8\3\2\2\2\u00cb"+
		"\25\3\2\2\2\u00cc\u00d0\7\21\2\2\u00cd\u00d0\7\22\2\2\u00ce\u00d0\5\30"+
		"\r\2\u00cf\u00cc\3\2\2\2\u00cf\u00cd\3\2\2\2\u00cf\u00ce\3\2\2\2\u00d0"+
		"\27\3\2\2\2\u00d1\u00d2\b\r\1\2\u00d2\u00d3\7\36\2\2\u00d3\u00da\5\30"+
		"\r\7\u00d4\u00d5\7\"\2\2\u00d5\u00d6\5\30\r\2\u00d6\u00d7\7#\2\2\u00d7"+
		"\u00da\3\2\2\2\u00d8\u00da\t\5\2\2\u00d9\u00d1\3\2\2\2\u00d9\u00d4\3\2"+
		"\2\2\u00d9\u00d8\3\2\2\2\u00da\u00e6\3\2\2\2\u00db\u00dc\f\6\2\2\u00dc"+
		"\u00dd\7 \2\2\u00dd\u00e5\5\30\r\7\u00de\u00df\f\5\2\2\u00df\u00e0\7\37"+
		"\2\2\u00e0\u00e5\5\30\r\6\u00e1\u00e2\f\4\2\2\u00e2\u00e3\7!\2\2\u00e3"+
		"\u00e5\5\30\r\5\u00e4\u00db\3\2\2\2\u00e4\u00de\3\2\2\2\u00e4\u00e1\3"+
		"\2\2\2\u00e5\u00e8\3\2\2\2\u00e6\u00e4\3\2\2\2\u00e6\u00e7\3\2\2\2\u00e7"+
		"\31\3\2\2\2\u00e8\u00e6\3\2\2\2\u00e9\u00ea\t\6\2\2\u00ea\33\3\2\2\2\u00eb"+
		"\u00ec\7\'\2\2\u00ec\35\3\2\2\2\32 \"+\60LWgmvz\u0086\u008d\u0093\u009b"+
		"\u00a2\u00a6\u00b8\u00c2\u00c4\u00ca\u00cf\u00d9\u00e4\u00e6";
	public static final ATN _ATN =
		new ATNDeserializer().deserialize(_serializedATN.toCharArray());
	static {
		_decisionToDFA = new DFA[_ATN.getNumberOfDecisions()];
		for (int i = 0; i < _ATN.getNumberOfDecisions(); i++) {
			_decisionToDFA[i] = new DFA(_ATN.getDecisionState(i), i);
		}
	}
}